Disable cross compiling. ./configure will try to execute a lot of runtime checks which are impossible during cross compiling. Resulting in the configure step failing.
Note that it would be possible to use conf_args.append(...) to define all those
required variables, but that's a brittle approach that needs to be fne tuned for
each target or the resulting binary will not be correct.

+++ recipes/apr/all/conanfile.py
@@ -65,6 +65,9 @@ class AprConan(ConanFile):
         if (self.settings.compiler == "apple-clang" and
             tools.Version(self.settings.compiler.version) == "12" and
             self.version == "1.7.0"):
+            if tools.cross_building(self):
+                raise ConanInvalidConfiguration("apr cannot be cross compiled due to runtime checks")
+
             with tools.chdir( self._source_subfolder ):
                 os.remove( "configure" )
                 self.run( "./buildconf" )
